REFACTORING: Rename files
- Rename sub8_utils/math_helpers -> msg_helpers, since every tool in
  there was a helper for rosmsg stuff
- Move simulate.py into nodes

Neither of these changes has any effect on the behavior of any code.
